move FakeSocket and FakeAPI shared classes to /test/utils
